let rocket StreamMap key-equal and hash use ebo
Summary: The Rocket client implementation keeps track of streams in a stream map, conveniently typed `StreamMap`. This is an F14 map with custom hash and key-equal types which do not actually have state. But the compiler thinks they have state and prevents the empty-base optimization. If an otherwise-empty class has a field of empty type, the containing class is not empty. We switch to an alternative technique to keep it empty.

Reviewed By: iahs

Differential Revision: D26906560

fbshipit-source-id: a76ddfce3ea8d9a3c1b6ac9884dcbe8fe05aadac