Consolidate stop processing cancelled request logic
Summary:
Prior to this change, a request was checked to see if it was still active three separate times before the request was handed off to user processing code. If not active, it would clean up the request object and return without sending a response.

After this change, this check only happens one time: right before user request handling code is executed.

Reviewed By: yfeldblum, praihan

Differential Revision: D27001435

fbshipit-source-id: cf2a0966d831e3fe8bda8552e2c6a24f43f77310