Streaming support in rust compiler
Summary:
Currently the rust thrift compiler does not support streaming of any kind. This diff adds support for methods that return a stream.

The implementation takes each stream response and treats it similarly to a normal response, using its own set of decoding functions.

One omission from this diff is that mocking stream methods is not yet supported. This is due to the stream trait not being cloneable

Reviewed By: Imxset21

Differential Revision: D26755112

fbshipit-source-id: 15cbe9d23f72c2325b2691a04f67602c0040450d