EnvelopeUtil: return PROTOCOL_TYPES v.s. int16_t
Summary:
The Rust client uses this. I'd like said client to get an actual protocol id so it can construct a `::apache::thrift::SerializedMessage` and call Thrift callbacks such as `onWriteData`.

The only values this ever returns are from PROTOCOL_TYPES, so rather than add a conversion in the Rust client, this diff updates `EnvelopeUtil` to just give us the variant we need.

Reviewed By: yfeldblum, Mizuchi

Differential Revision: D27088186

fbshipit-source-id: 808a8689a63c18543bff6c5f901a6b6604fe3c3e