Add is_eligible_for_constexpr for Thrift types
Summary: Add the `is_eligible_for_constexpr` class that determines if the operations on the C++ representation of a Thrift type can be constexpr and, in particular, whether we can define the move constructor `constexpr T(T&&) = default`. This will allow us to outline large move constructors while keeping small constexpr and noexcept ones defaulted. Currently we rely on users doing this manually via the `cpp.noexcept_move_ctor` annotation which is error-prone and not scalable. We would like to outline many constructors automatically for build speed reasons.

Reviewed By: yfeldblum

Differential Revision: D27136295

fbshipit-source-id: 52f3248155e009a16e45ef3c1d2befc1cdd331d7