Log count of active requests at 10 Msec intervals to thriftDbg
Summary: Similar to recent_request_counts that shows count of arrived requests in 10 msec intervals from last 5 seconds, this change adds tracking for count of in-flight requests. SR uses active+pending requests for load balancing, so this new data will help us understand how good we are balancing the load to the server.

Reviewed By: sethdelliott

Differential Revision: D27225624

fbshipit-source-id: e2fc9aa37fb4b380d7921fe75d52d7bb55e57b7b