Use typesafe equals
Summary:
Our Hack codebase is moving to use the "identical" checks, `===` and `!==`, when possible. This avoids legacy PHPism conversions that are unintuitive and surprising to new developers.

This change is **100% safe** because it can be proven to never change behavior; it's main benefit is reducing the number of extant `==`/`!=` checks in our codebase.

Reviewed By: DavidSnider

Differential Revision: D27188599

fbshipit-source-id: ed8c3603b8ef20a5b5693d4c19e8570a12125d97