Introduce `ThriftServer::setQuickExitOnShutdownTimeout` to quick_exit the server instead of aborting
Summary:
The default behavior of a `ThriftServer` is to abort if the server fails to drain
its active requests after a timeout value.
Most of the time, this type of problem happens with SRProxy (old) sidecar. When a
sidecar is shutdown it drains the Thrift server and abort if it can't succeed
after a timeout.

Most users don't care about this, and recently it had been so
oblivious that some teams exceeded their coredumper capacity.

Reviewed By: yfeldblum

Differential Revision: D27202218

fbshipit-source-id: 02a057b7dcd9a691ed83db47368308db18334997