Fix race condition in FunctionRunner.
Summary:
Right now when a FunctionRunner is stop()ed and on its exiting path for repeated functions, it can potentially acquire back a destroyed mutex, because stop() doesn't wait for run() to finish in this case.

One way is to add a baton to wait for run() to exit and then proceed with destruction in this case.

Reviewed By: yfeldblum

Differential Revision: D27495599

fbshipit-source-id: e91de4f522be0ef7658450f517913a889f92c4e7