client request buffer for version dependent frames
Summary:
With version 7, rocket client and server are switching to metadata push frames instead of extension frames for headers push and interaction termination, so long as the other side supports it. On the client side, the server's setup response may not have been received yet when a headers push or interaction termination is initiated, causing the wrong type of frame to be sent.

This change adds deferred initialization to RequestContext objects, so they can be properly initialized at a later point when the server version is available. When such an uninitialized RequestContext is scheduled for write, RequestContextQueue will start buffering it and all subsequent contexts until the server version is received and the buffer can be resolved. This ensures that the intended order of writes is kept intact.

Since rocket server only responds to setup frames starting with version 5, this change also sets the minVersion required by the client. Additionally, a timeout of 10s is added to close the connection in case the setup response is not received.

Reviewed By: iahs

Differential Revision: D27252879

fbshipit-source-id: db69d98d564a3ebbfaa53ad5e8328b307735d189