replace extension frames 0xFB0, 0xFB2 with metadata push frame
Summary:
Replace extension frames 0xFB0 and 0xFB2, which are used for stream headers push and interaction termination respectively, with metadata push frames. Both client and server will still accept the extension frames as before, and the metadata push frames are only sent if the other side is known to support it (i.e. version >= 7).

In some cases the client may want to send interaction termination or headers push before the server's setup response has been received and thus the server version is not known yet. In that case, the server version is assumed to be 0 and the extension frames are sent (which the server will still accept as before). This issue is addressed in a follow up change.

Reviewed By: iahs

Differential Revision: D26983407

fbshipit-source-id: a5bbbf1786a8a7c2f24537866234c7ab60047755