Fix ThreadManager::Observer data race when setting and running observer at the same time.
Summary:
Right now we're holding a static ThreadManager observer in ThreadManager.cpp, with every implementation sharing the same instance. There's a shared mutex around it to protect write operations, but in the read path we just read the pointer without synchronization.

This diff firstly migrates the mutex + object pair from class statics to a singleton in anonymous namespace, so that we're more clear which observer we're referring to in the code. Also we use an atomic boolean in the read path to preserve the original optimization without data race.

Reviewed By: yfeldblum

Differential Revision: D27498547

fbshipit-source-id: 7414fb85db00a7a114720515d621285bfa67f099