Fix failure in InstrumentationTest.
Summary:
Since ScopedServerThreadInterface is running in a separate thread, we need to wait for the servers to start to get the actual number of the server, otherwise we have a race condition.

The simplest way is probably wait for the first request to come back so that we know the server serves.

Reviewed By: sethdelliott, praihan

Differential Revision: D27578563

fbshipit-source-id: 9916c29e2129bcd556c011c876ef8b1ed06edb7b