Add an error message on the pyname field for metadata if unstructured annotations not enabled
Summary:
Unstructured annotations are currently gated behind a thrift cpp2 option to avoid unnecessary code generation, and it might be confusing for users to have to know this, so this error should help users know to enable it.

If unstructured annotations aren't enabled, the unstructured_annotations field on the metadata comes back as None, so this just checks whether it's none or an empty dict.

Reviewed By: junzh0u

Differential Revision: D27571584

fbshipit-source-id: 97720502b30aea4c429bf5fcd69089fd9272f72d