work around clang-9 ASAN ICE
Summary:
Some projects still build build with clang-9. For that version of clang, coroutines and ASAN are a volatile mixture. Merely accessing a data member from a coroutine member function causes the compiler to crash.

We can work around the problem trivially by replacing the member function with a static member and passing in `*this`.

Reviewed By: yfeldblum, ispeters

Differential Revision: D27688850

fbshipit-source-id: 5c69b927ba7a6a5bfe9dbe03c564d3c90191ecb7