Update HeaderClientChannel constructor to take transport unique_ptr
Summary:
HeaderClientChannel rocket upgrade logic accepts a shared_ptr and then makes a dynamic assumption that the shared_ptr has a specific deleter. This assumption is not true even for some unit tests in thrift/.
Instead we'll have to update HeaderClientChannel constructor to accept a unique_ptr. As a follow up it will become possible to move ReleasableAsyncTransportPtr into HeaderClientChannel.cpp so that AsyncSocket and other AsyncTransports don't have to use it.

Reviewed By: yfeldblum

Differential Revision: D27672171

fbshipit-source-id: c3acb863b72001089012178374df45cee51f7638