ThriftServer snapshot should cointain client_id
Summary:
In Rocket, we pass client_id as a dedicated field in RPC metadata. As a result it does not appear in the regular key/value headers map. So we were missing this header when snapshotting requests on the server.

Another implication is that Cpp2RequestContext's headers map is also "fixed up" - we explicitly insert this header into the map<string, string> .... which is suboptimal but is currently needed for legacy reasons. This diff introduces `clientId` API on `Cpp2RequestContext` allowing to fetch `clientId` directly and more efficiently than searching the the map.

`ProcessParams` datastruct also exposes `clientId` API to allow for more efficient access to the data.

Reviewed By: andriigrynenko

Differential Revision: D27698667

fbshipit-source-id: 081c25fe17bccce42a9da8fba2c655f000bec90b