Add missing dependency on RocketUpgrade headers to thriftcpp2 target (#422)
Summary:
One source file in this target, async/HeaderClientChannel.cpp,
depends on the generated header RocketUpgradeAsyncClient.h, so it
needs to depend on the target that generates this header.

This causes a build error with samurai due to an incorrect build
order, and can be reproduced with ninja as well by building
thrift/lib/cpp2/CMakeFiles/thriftcpp2.dir/async/HeaderClientChannel.cpp.o
directly with an empty .ninja_deps.

Pull Request resolved: https://github.com/facebook/fbthrift/pull/422

Reviewed By: iahs

Differential Revision: D27694167

Pulled By: yfeldblum

fbshipit-source-id: 0a59086af6ed3caa52cbbe5761e8552d5f2d0a85