Thrift Interactions support for py3 clients
Summary:
This is an implementation of the cpp2 interactions in python, implemented via a cython wrapper.

The way interactions are handled, step by step:
* First in the cython `create{{service:name}}` function, a future gets started and run with the c++ client wrapper interaction creation function (which could actually be done sync but needs to be run in the EventBase thread for thread safety)
* Then, in the c++ client wrapper function `ClientWrapper::create{{service:name}}` the c++ interactions creation code gets called, and this is done in the EventBase thread.
* Then after that the client functions identically to a standard client under the hood, with only a few types changed since it's an interactions client and not a regular client

One additional note that I moved some code in `Callbacks.mustache` and some code in `clients_wrapper.cpp` out into separate mustache files to avoid duplicating mustache code between a regular client and an interactions client, and I had to move 1 line of code in the client wrapper function call code to make it refactor out cleanly, so a bunch of the fixtures have 1 line of code added 1 line deleted.

Reviewed By: nanshu

Differential Revision: D26701079

fbshipit-source-id: cecb4f2638dc1f714cba08c2a055a2ccf90ddc5a