Factor out field serialization logic
Summary:
To keep the templates simple, I've removed an optimization I added a couple diffs back, that removed double nullptr checks when a field is both a cpp.ref and has terse_writes enabled (i've kept the optimization for all other cases). This also makes the name of the binding used to removing the double checks much more straightforward.

The problem with keeping this logic is an odd interaction between terse_writes, cpp.ref, and unions. Ideally, union fields would have both 'field:optional?' and 'field:terse_writes?' be false. Unfortunately, when terse_writes is provided on the command line, it is true for all union fields (except shared references for some reason), and the c++ field object does not know what kind of struct owns it, so it can't special case 'field:must_write_cpp_ref?' for union fields.

Reviewed By: Mizuchi

Differential Revision: D27717452

fbshipit-source-id: d1665d1dd222acc7a460e5acfa3fbf6bd58cf26c