makeTestClient should use global CPU executor as callback executor
Summary:
If clients are constructed using `makeTestClient`, and the user calls `future_*` methods
on the returned clients, `PooledRequestChannel` will try to use the callback executor to
run the callback chained onto the returned `future_` call.

Previously, this callback was created with nullptr, resulting in failures when
`PooledRequestChannel` tried to use the callback executor.

This change makes `makeTestClient` always use the global (immutable) CPU executor
as the callback executor for `future_*` methods.

Reviewed By: iahs

Differential Revision: D27773514

fbshipit-source-id: a8b9711e6949e7c9f5468348f814b3f57fd0b2b4