Fix thrift Go codegen for typedefs
Summary:
Found an error coming from `t_go_generator.cc` (the script that performs codegen from thrift -> Go), because it didn't handle thrift `typedef`s properly in many cases.

This diff should fix it.

Instead of trying to refer all the way to the very base module for all purposes, it does this only to check if the actual Thrift struct is defined as a union, etc. For anything else, only the module/package one level below is used, where the original struct might exist in `typedef`-ed form.

This should now enable use-cases like this:
```
In A.thrift:

# define SomeStruct

In B.thrift:

# import A.thrift
# typedef A.SomeStruct SomeOtherStruct

In C.thrift:

# import B.thrift, but not A.thrift
# use B.SomeOtherStruct here
```
to be correctly translated to Go (`typedef`s over multiple levels).

Reviewed By: dlespiau

Differential Revision: D27783113

fbshipit-source-id: 7c09f5a10f3fa1e560091d1a5168e165c377d7ca