AnyRegistry::load directly into the passed object
Summary:
Instead of deserializing into a default-constructed object and then assigning it to the passed reference, deserialize directly into the passed reference.

This changes the semantics, as it may retain some previously set fields while the existing behavior is to always start from a default-constructed object. Note that this is consistent with the standard Thrift deserialization API.

Reviewed By: luciang

Differential Revision: D27809703

fbshipit-source-id: cf8f22ab6cec9eece2291576fedf59e8d2df670f