Move field template logic to shared folder.
Summary: I would have merged common/field_value and a few other things into field/*, but that would add `this->` in a bunch of new places, which is used to avoid potential collisions with local variable names. I will update those in a different diff. Note that the ideal solution to this would be to rename the struct fields to include the `_` suffix, but we can't do that until all of the raw accesses have been removed.

Reviewed By: Mizuchi

Differential Revision: D27611923

fbshipit-source-id: b50c27efa6838149f59a0b62bb654301aeb16c95