generate AsExceptionType traits and impls for services
Summary:
Often, people want to check if an exception from a client for a specific
service is a certain exception type.

instead of making people keep these if Some chains up to date with all the
functions in a thrift service, we can instead generate a trait that represents
the unpacking they want. In addition, we can implement this trait not only for
the Error types return'd by the methods themselves, but also conveniently
search an anyhow chain for that specific error

unfortunately, this relationship is not yet represented in the mstch types very
well, so I had to generate a new structure to do this

Reviewed By: dtolnay

Differential Revision: D27140374

fbshipit-source-id: 5d58a10f1d65ea44e8caa54d97c5a6c7a0d892b2