Implement py.adapter annotation
Summary:
Only pure-python codegen for now, fastproto and type stubs will be subsequent diffs.

Same design as for `hack.adapter`: if an annotation is present, call a static method to convert from thrift type to runtime type during deserialization and vice versa.

Only works for structs for now to simplify the implementation, could be extended to other types if the future if necessary.

Reviewed By: ot, junzh0u

Differential Revision: D27664099

fbshipit-source-id: 45f7c299315578d65ed1f58ba5dfe8951373f2b2