thrift: support /// doc comments
Summary:
This adds support for `///` as a doc comment in .thrift files. These
may either be a single line or multiple successive `///` lines (possibly with
intervening blank lines).

From a survey of existing .thrift files, there seem to be two cases - one is
using `/////...` for banner comments, and the other are comments which appear
to be doc comments (ie, the authors already expected /// to be supported as a
doc comment syntax).

Consequently, this only treats `///` followed by a non-`/` as a doc comment.

Most of the complexity of the diff is in `parsing_driver::clean_up_doctext` and
a corresponding change in the lexer to recognize /// comments as doc comments,
and do appropriate prefix/suffix stripping.

#forcetdhashing

Reviewed By: yfeldblum

Differential Revision: D27974242

fbshipit-source-id: 2ceb06468806e3070689ea7be0f09f612c497195