Byte-oriented backpressure for streams
Summary:
Streams continuously generate streaming responses, so long as they have credits left. Network congestion can lead to these generated payloads accumulating in the server send buffer, potentially causing OOMs when there are many parallel streams and/or large payloads being generated.

This change adds BaseThriftServer config parameters `egressBufferBackpressureThreshold` and `egressBufferRecoveryFactor`. When set, AsyncSocket's buffer callback is used to check the size of outstanding writes whenever it needs to buffer a payload. If the configured threshold is exceeded, all streams for the connection are paused until load falls below the recovery factor of the threshold (or the socket buffer is cleared entirely).

Reviewed By: iahs

Differential Revision: D27957403

fbshipit-source-id: 194e0e4a718ca25a868daf4c55fe607611b22318