thrift: Add boxed_value_ptr class
Summary:
Instead of using a raw std::unique_ptr as the storage for
optional_boxed_field_ref, let's wrap it in a custom type so we have
the flexibility to customize behaviour a little bit if needed. This
avoids having to fiddle too much with the mustache templates. Instead,
we can make additions to the wrapper type instead which is a lot saner
to write and test code for.

Reviewed By: Mizuchi, vitaut

Differential Revision: D27896185

fbshipit-source-id: 94a464c4e9d8ad90e37f81de5d7486b854eb1b0e