thrift: Add optional_boxed_field_ref
Summary:
This will be the field_ref implementation for
accessing fields marked with the new cpp.recursive annotation.

One oddity is that the template parameter is not the type we
store in the boxed_value_ptr but the boxed_value_ptr type itself
so we can do type shenanigans to copy the ref and constness from
the boxed_value_ptr type to the contained type which is necessary
to provide the same API as optional_field_ref.

Tests are the same tests as optional_field_ref. One small difference
is that we don't initialize the boxed field ref name_ field in
TestStructBoxedValuePtr with the "default" string because this
makes the field ref non-empty while the tests assume it is empty so
we don't initialize it instead. This allows reusing the same tests
for both optional_field_ref and optional_boxed_field_ref. Reusing
the same tests is useful for ensuring the APIs don't diverge.

The move_from test has a small addition so we can specify what the
moved from state is which differs between optional_field_ref and
optional_boxed_field_ref.

Reviewed By: Mizuchi, vitaut

Differential Revision: D27896186

fbshipit-source-id: 0b966a58309338585de5cd2491fa767d6a840a18