thrift: Add new cpp.box annotation
Summary:
By default, thrift structs nested in other thrift structs are
embedded into the parent struct. This poses problems when a
parent struct embeds many other structs which are unused most
of the time. An example of this is SkynetStructs.thrift. Because
the structs are embedded without any indirection, the empty unused
structs have an adverse effect on the size of the parent struct in
memory. Due to the amount of embedded structs, Skynet's structs
balloon to huge sizes, even though most of the time, the embedded
structs are completely empty.

This issue is the cause of a performance degradation when trying
to move to thrift-py3 from thrift-py2 for code using the Skynet
thrifts (see https://fb.workplace.com/groups/thrift.py3/permalink/1863906357094481/).

The first solution to the problem I came up with was to a layer of
indirection using cpp.ref. This turns all the embedded thrift structs
into std::unique_ptr's. A struct embedded in another struct this way
only takes up a single extra pointer worth of space instead of adding
the size of an entire struct to the parent struct's size. If the field
is not set, this unique_ptr never allocates which saves a lot of memory
overhead. Applying this solution solves the performance degradation
seen in the linked post.

Unfortunately, Adding the cpp.ref annotation changes the API of the
C++ generated thrift code. There's already a lot of code using the
thrift generated C++ skynet code which means changing the thrifts
to use cpp.ref for embedded structs is a breaking change and would
require changing all callers along with adding the cpp.ref annotation.
This extra work makes testing out the fix prohibitve, as migrating all
the calling code would take too much time and be prone to errors as
well.

To fix this, we need an annotation that tells thrift to add a layer
of indirection internally (by storing the embedded struct as a pointer
allocated on the heap) without changing the public API of the C++
generated code. This diff adds exactly such an annotation which we
name cpp.box.

The annotation can also be used used as a replacement for cpp.ref
with the added advantage of not changing the public API of the
generated C++ code (too much).

As it is a new API, the annotation does not support any deprecated
constructs so users will have to remove any usage of deprecated
features from their code and user's code before they can enable this
new annotation.

For now, we only implement the annotation for struct fields but in
the future, union support can be added as well.

The changes in this diff add the new annotation and integrate it into
the code generate using the boxed_value_ptr and optional_boxed_field_ref
classes that were added in previous diffs.

#forcetdhashing

(Note: this ignores all push blocking failures!)

Reviewed By: Mizuchi

Differential Revision: D27628435

fbshipit-source-id: 260f585daea4947742a83ff3456b2633a5bd2f8c