Propagate `cpp2.noncopyable` annotations up to the parents of fields.
Summary:
This diff is in preparation for defining the copy constructors out-of-line in D27817371. Previously, if a struct `S` was marked with the `cpp.noncopyable` annotation, another struct `W` that wraps `S` would also be noncopyable due to the `= default;` definition which deduces copyability.

In D27817371 we lose the ability to deduce copyability since the copy constructor would only be declared in the header and defined in the source file. This patch deduces copyability at the code generation level by considering structs such as `W` noncopyable if it transitively contains a noncopyable type such as `S`.

Note: `W` can still be copyable while containing `S` if `S` is contained via `std::shared_ptr`. For example:
```
struct S {} (cpp.noncopyable);

struct W {
  S s (cpp.ref_type = "shared");
}
```

Reviewed By: vitaut

Differential Revision: D28079518

fbshipit-source-id: 80e22112270328c1d454b32607839a025b443d6e