Avoid destroying interaction while it still has a server stream AsyncGenerator
Summary:
Accessing the interaction object from inside a SCOPE_EXIT in stream generator may segfault. This diff fixes it.

Context: such SCOPE_EXIT is kind of the only way of getting notified when the stream is closed, so I'd like to be able to do things there. Our specific use case is an interaction having one sink and one stream, and when either of them closes we want to close the other one too. So in stream's SCOPE_EXIT we set some state to advise the corresponding sink to close too.

Reviewed By: iahs

Differential Revision: D28106123

fbshipit-source-id: 884311954d7d60fe529698f3a8c973892378389f