Avoid move/copy construction of parameters in sendRequestAsync
Summary:
Currently, `maybeRunInEb` takes a lambda which must capture (either by copy or move) the various
parameters to `sendRequestAsync`. (The most expensive of these captures is the copy of `RpcOptions`, but
they all sort of add up in cost, to around a few percent of the overall cost of `sendRequestAsync`.) This
diff avoids the copy/move capture cost in case the channel send function is called inline.

Reviewed By: yfeldblum

Differential Revision: D28136505

fbshipit-source-id: 9fee91cace15e33f53043f92e0b181e41201d70f