rust: generate ::new methods for newtypes
Summary:
These new methods can be useful when referring to generated newtypes in rust macros. Specifically, if macro's fragment-specifier is a Type (`:ty`), the macro won't allow you to construct a new instance of this type with parenthesis, like this:
```
macro_rules! bla {
  ($a: ty) => {
    fn create() -> $a {
      $a(0)
    }
  }
}
```

You can's define methods on foreign types, and you can only implement local traits on foreign types. So in cases when macro itself needs to be imported by other traits, you are out of luck.

This change makes it possible to do the following for thrift-generated rust newtypes

```
macro_rules! bla {
  ($a: ty) => {
    fn create() -> $a {
      $a::new(0)
    }
  }
}
```

Reviewed By: krallin

Differential Revision: D28253734

fbshipit-source-id: dbfd4b507d9eec72542e130056d6c243f6f976c6