Expose setDefaultCompressionConfig on RocketClientChannel
Summary: mcrouter uses generated Thrift clients directly over `RocketClientChannel` and needs support for compression. This diff exposes the ability to set compression directly on `RocketClientChannel`, which matches a recent change to `HeaderClientChannel` and avoids the need for mcrouter to maintain an extra channel (similar to the now-killed `RocketTestClientChannel`) on top.

Reviewed By: iahs

Differential Revision: D28258756

fbshipit-source-id: 18afda628699a287222bc4cb0990a5540721b5bf