Fix a bug in structured annotation type tags
Summary:
The if block that was testing the type tags was just testing `if [ENUM]` which would just always return true if the enum being tested against wasn't 0, which meant that some basic types were all just getting tagged as boolean instead of their real types.

This is just setting that test to `if value is [ENUM]`.

Reviewed By: yfeldblum

Differential Revision: D28269667

fbshipit-source-id: eae2abcd448735e360a4d15cf9850601f99a5ac3