add ErrorResult interface for result structs
Summary: When thrift servers run Run or RunContext, the caller has no way of generically knowing if the result contains a structured exception. The results now implement the `WritableResult` interface, which allows the caller to extract an exception. The signature of Run and RunContext are not changed, as this would break callers using the public `ProcessContext`. Unfortunately, servers will need to assert the interface to get the error with this implementation.

Reviewed By: codyohl

Differential Revision: D28295454

fbshipit-source-id: fd7c51a507db582504822051e57e031dc6d8c0d3