Release shared_ptr<THeader> before potentially suspending fiber
Summary:
This diff ensures `RocketClientChannel::send*` does not artificially extend the lifetime of the `shared_ptr<THeader>` across a fiber suspension point. Note that this should be safe from the point of view that `ContextStack` handlers cannot access the request's `THeader` after `clientSendT` returns; nobody should be circumventing this intention in the case of directly using RCC with fibers.

#forcetdhashing
#buildmore

Reviewed By: yfeldblum

Differential Revision: D28204400

fbshipit-source-id: 909544090dca2445a4bfbb09d1c43f189bb6efc5