Don't throw for missing fields
Summary:
`vars` is used to find values in seeds for the older python thrift implementation. That builtin does weird things with double underscores. With that further fuzzing won't be able to fuzz that field but will try. This results in lots of key errors and attribute errors.

This adds the appropriate safe guards to make sure that missing fields are just skipped and don't error out.

Reviewed By: yfeldblum, jermenkoo

Differential Revision: D27516087

fbshipit-source-id: 8f3d57f70329ae6c34762bed95cc8618f0e10d7f