add wrapper of ReadState to handle indexing
Summary:
Currently, there is small regression for index related logic, even when index is not enabled. The regression comes from
* Check whether index exists. Even with prefect branch prediction there is still overhead
* Initialize index related state.

This diff moved the logic out of ReadState, into `ProtocolReaderStructReadStateWithIndex` which is wrapper of ReadState for indexing related logic. The wrapper doesn't change API of ReadState, but only added indexing related logic. This version will only be used if indexing is enabled.

Reviewed By: vitaut

Differential Revision: D28322228

fbshipit-source-id: c68c26cd86a16d264a6566f38722792301de6d5a