Java: Skip union field that are marked-as-set but equal to `null`
Summary:
The new serialization mechanism throws an exception when trying to serialize a marked-as-set but `null` union field while the previous one (bytecode generated) was just skipping the field.

This diff change the behavior of the new code to match the previous one.

In general, `null` values are not serializable for non-optional fields. So we should either forbid them or skip them. Since non-primitive fields are defaulted to `null`, the only viable solution is to silently ignore `null` valued field.

#forcetdhashing

Reviewed By: avalonalex

Differential Revision: D28605409

fbshipit-source-id: 8ee70c21d484bfd222dd1219a0c24ef4c6df565e