Split out t_placeholder_typedef and add t_type_ref::deref.
Summary:
Simplifying the t_typedef class and eliminating a lot of unsafe accesses that assumes both the t_type_ref has been initialized properly, and that all placeholder typedefs have been resolved. If either was not the case, a nullptr access would occur. Now an exception is thrown instead.

This is also another step towards making t_type_ref a real t_node (by isolating one of the use cases it will need to absorb).

t_type_ref::deref also skips over the t_placeholder_typedefs. There are a lot of bugs related to this in the compiler, where either the placeholder isn't traversed properly or all typedefs are ignored. Ultimately, the right way to fix these bugs is to merge t_type_ref and t_placeholder_typdef, but this is non-trivial and the deref API solves the problem for any case that doesn't care about unstructured annotations (e.g. everything but the generators). I've left the code with these bugs alone, but they are now all accessing the t_type via deprecated APIs, so they are easy to identify, and will be fixed when the access is upgraded.

Reviewed By: vitaut

Differential Revision: D28594298

fbshipit-source-id: fb2b722582934a3e3094f8b00a69505e98a1e3b0