Optimize RocketClientChannel
Summary:
Currently, RocketClientChannel are two separate allocations and the fields inside of RocketClient are mostly randomly distributed, which is not very cache friendly and leads to us having lots of cache misses along the request processing path.

This diff changes `RocketClientChannel` to inherit from `RocketClient` and reorders fields inside of `RocketClient` to have fields on that are always accessed on request path to be collocated at the beginning of `RocketClient` object.

Reviewed By: iahs

Differential Revision: D28595434

fbshipit-source-id: d318089712fe76c89cb6531395043454e13d61aa