Fix RequestInstrumentationTest.ConnectionSnapshotsTest
Summary:
Test was timing out and quickstack shows that sometimes requests can have a spurious `TTransportException`. This means that `waitForRequestsThenSnapshot` blocks forever since the target count will never be reached.

I noticed during a test run that each iteration would get slower as time went on. I suspect the next iteration of the test starts as soon as `TearDown` ends rather than when the server is destroyed.

The solution is to block until the server is done working. Blocking until `ThriftServer::getActiveRequests() == 0` also fixes a race condition for checking that connections snapshot is empty (since we don't block on the RPC `SemiFuture`s).

Differential Revision: D28707379

fbshipit-source-id: 987175870eedc53d6a50c6a843f65aa2399a138a