refactor includeInRecentRequestsCount
Summary:
The main goal of this diff was to move `includeInRecentRequestsCount` from `RequestsRegistry::DebugStub` to `RequestStateMachine`. We are going to reuse this flag in some decisions made for adative concurrency and hence are moving this flag into request's data.
So now, `RequestStateMachine`'s ctor accepts this parameter.

This however, caused a few other changes. Notably, we need to know the name of the request when we construct `RequestStateMachine`. Notice that this thing is constructed right after `Cpp2RequestContext` is constructed and ideally we should just get method's name from the latter.

However, `Cpp2RequestContext` construcutor did not accept method's name, instead it was later provided to it using `Cpp2Request::setMethodName` API. Oh, and for Rocket we were calling `setMethodName` twice, so big mess.

Provided all the background, here is the set of changes:
* moved  `includeInRecentRequestsCount` from `RequestsRegistry::DebugStub` to `RequestStateMachine`
* added `methodName` to ctor of `Cpp2RequestContext`
* removed  `Cpp2RequestContext::setMethodName` (and fixed the few callsites that were using this API)
* this all results now in the correct computation of `includeInRecentRequestsCount` for Header, so adjusted the unit test accordingly.

Reviewed By: praihan

Differential Revision: D28577720

fbshipit-source-id: f83d927dcbd80bf3a77359cdbf4bb957d72389fc