Provide extension points for hashable structs
Summary: This makes it possible to use arbitrary structures as keys in all Frozen associative structures by specializing the hash methods in the type's layout as show in the included unit tests. Note that hash specializations must emit the same hash value for the thawed and frozen representations of the same value.

Reviewed By: vitaut

Differential Revision: D28839759

fbshipit-source-id: e2762950990b11d3ecf8d0f11767f1a7fee8ef51