Adaptive concurrency controller
Summary:
Adaptive concurrency controller. It is not hooked up to Thrift Server itself, but subsequent diff would fix that.
This piece is basically just and algorithm + unit tests.
The main idea of adaptive concurrency controller is that the algorithm will try to find `max_requests` parameter for Thrift Server so that admitted request demonstrate certain latency characteristics.

The algorithm itself is described in the comments, and since we have not invented it but rather adopting the previously published approach, the original references to the algo are given in the comments as well.

Our controller receives its configuration via an Observer. Notice that some parameters to the algo are hard coded (such as 500ms resampling interval and 5min recalibration, some dampening values and so on). This is intentional as we would like to minimize number of knobs give to our users for the simplicity

The controller exposes `requestStarted`/`requestFinished` APIs that the consumer (i.e. ThriftServer) will invoke on every request it wants to observer for the purpose of adaptive concurrency controller.

Reviewed By: praihan

Differential Revision: D28333961

fbshipit-source-id: 16ed6b3e73e7095ad2f64e9f19e4d72cdcacf9f0