Record last deserialized protocol
Summary:
Currently for lazy deserialization, we hard-coded Compact Protocol everywhere since it's the only supported protocol.

In order to support other protocols in the future, we need to record what's the protocol that's used during last deserialization, so that when we deserialize lazy field on the fly, we can use the correct protocol, rather than just Compact Protocol.

Reviewed By: yfeldblum

Differential Revision: D28927522

fbshipit-source-id: 66dd8c1b612317aa82d8b7d1018146477280464e