Fix data race in stopAccepting()
Summary:
This is fairly sloppy, but should fix the threading issues without introducing
more race conditions or deadlock. Since there are two locations where
`tlsCredProcessor_` is lazily constructed which happen in different threads,
it's possible for a race at construction (although I don't know if we've ever
seen this?).

The destruction race we believe we have seen, as it occurs if `stopListening()`
is called, since it calls `stopAcceptingAndJoinOutstandingRequests()`, which
(prior to this diff) destroyed the `sharedSSLContextManager_` prior to the
thrift server destroying itself. Since the context manager manages the ticket
key logic which the `tlsCredProcessor_` member of the thrift server modifies
(via its own thread), this can cause a crash.

I don't imagine we'll see contention here since these only get modified during
secret updates. The secrets themselves are stored outside the cred processor
class, meaning acceptors can continue working as expected while the write lock
on the processor is held.

Reviewed By: mingtaoy

Differential Revision: D27981839

fbshipit-source-id: 1a864cd112e9595b9500639bbac676f745cd2579